chore: separate useIntersectionObserver
